Introduce the AI_Conditional and Route_Permission_Trait
