+++ src/ai-generator/user-interface/get-usage-route.php
@@ -8,7 +8,7 @@ use Yoast\WP\SEO\AI_Generator\Application\Token_Manager;
 use Yoast\WP\SEO\AI_Generator\Domain\Exceptions\Remote_Request_Exception;
 use Yoast\WP\SEO\AI_Generator\Domain\Exceptions\WP_Request_Exception;
 use Yoast\WP\SEO\AI_Generator\Domain\Request;
-use Yoast\WP\SEO\Conditionals\No_Conditionals;
+use Yoast\WP\SEO\Conditionals\AI_Conditional;
 use Yoast\WP\SEO\Main;
 use Yoast\WP\SEO\Routes\Route_Interface;
 
@@ -21,7 +21,7 @@ use Yoast\WP\SEO\Routes\Route_Interface;
  */
 class Get_Usage_Route implements Route_Interface {
 
-	use No_Conditionals;
+	use Route_Permission_Trait;
 
 		/**
 		 *  The namespace for this route.
@@ -50,6 +50,15 @@ class Get_Usage_Route implements Route_Interface {
 	 * @var Request_Handler
 	 */
 	private $request_handler;
+
+	/**
+	 * Returns the conditionals based in which this loadable should be active.
+	 *
+	 * @return array<string> The conditionals.
+	 */
+	public static function get_conditionals() {
+		return [ AI_Conditional::class ];
+	}
 
 	/**
 	 * Class constructor.
@@ -103,21 +112,5 @@ class Get_Usage_Route implements Route_Interface {
 		}
 
 		return new WP_REST_Response( $data );
-	}
-
-	/**
-	 * Checks:
-	 * - if the user is logged
-	 * - if the user can edit posts
-	 *
-	 * @return bool Whether the user is logged in, can edit posts and the feature is active.
-	 */
-	public function check_permissions(): bool {
-		$user = \wp_get_current_user();
-		if ( $user === null || $user->ID < 1 ) {
-			return false;
-		}
-
-		return \user_can( $user, 'edit_posts' );
 	}
 }
